FIX #18805

+++ htdocs/install/mysql/tables/llx_product.sql
@@ -47,11 +47,11 @@ create table llx_product
   price_base_type               varchar(3)   DEFAULT 'HT',
   cost_price                    double(24,8) DEFAULT NULL,          -- Cost price without tax. Can be used for margin calculation.
   default_vat_code              varchar(10),                        -- Same code than into table llx_c_tva (but no constraints). Should be used in priority to find default vat, npr, localtaxes for product.
-  tva_tx                        double(6,3),                        -- Default VAT rate of product
+  tva_tx                        double(7,4),                        -- Default VAT rate of product
   recuperableonly               integer NOT NULL DEFAULT '0',       -- French NPR VAT
-  localtax1_tx                  double(6,3)  DEFAULT 0,
+  localtax1_tx                  double(7,4)  DEFAULT 0,
   localtax1_type                varchar(10)  NOT NULL DEFAULT '0',
-  localtax2_tx                  double(6,3)  DEFAULT 0,
+  localtax2_tx                  double(7,4)  DEFAULT 0,
   localtax2_type                varchar(10)  NOT NULL DEFAULT '0',
   fk_user_author                integer DEFAULT NULL,               -- user making creation
   fk_user_modif                 integer,                            -- user making last change
