+++ htdocs/install/mysql/tables/llx_product_fournisseur_price.sql
@@ -39,10 +39,10 @@ create table llx_product_fournisseur_price
   default_vat_code	    varchar(10),
   barcode                       varchar(180) DEFAULT NULL,          -- barcode
   fk_barcode_type               integer      DEFAULT NULL,          -- barcode type
-  tva_tx				double(6,3) NOT NULL,
-  localtax1_tx		    double(6,3) DEFAULT 0,
+  tva_tx				double(7,4) NOT NULL,
+  localtax1_tx		    double(7,4) DEFAULT 0,
   localtax1_type        varchar(10)  NOT NULL DEFAULT '0',
-  localtax2_tx		    double(6,3) DEFAULT 0,
+  localtax2_tx		    double(7,4) DEFAULT 0,
   localtax2_type        varchar(10)  NOT NULL DEFAULT '0',
   info_bits				integer NOT NULL DEFAULT 0,
   fk_user				integer,
